Minor fixes

+++ src/main/scala/com/mehulpoddar/stockexchange/models/Player.scala
@@ -1,15 +1,15 @@
 package com.mehulpoddar.stockexchange.models
 
 case class Player(name: String,
-                  cashInHand: Double,
+                  cashInHand: Int,
                   allottedShares: Map[String, Int],
-                  cards: Map[String, Seq[Double]] = Map.empty[String, Seq[Double]]) {
+                  cards: Map[String, Seq[Int]] = Map.empty[String, Seq[Int]]) {
 
   override def toString = s"${name}\nCash: $cashInHand\nShares: ${allottedShares.mkString(", ")}\n"
 
   def organizeCards(playerCards: Seq[Card]): Player = {
-    val newCards = playerCards.foldLeft(Map.empty[String, Seq[Double]])((map, card) => {
-      val mapValue = map.getOrElse(card.companyCode, Seq.empty[Double])
+    val newCards = playerCards.foldLeft(Map.empty[String, Seq[Int]])((map, card) => {
+      val mapValue = map.getOrElse(card.companyCode, Seq.empty[Int])
       map + (card.companyCode -> (mapValue ++ Seq(card.value)))
     })
     copy(cards = newCards)
@@ -31,7 +31,7 @@ case class Player(name: String,
     copy(cashInHand = updatedCash, allottedShares = finalUpdatedShares)
   }
 
-  def netWorth(companies: Map[String, Company]): Double =
+  def netWorth(companies: Map[String, Company]): Int =
      allottedShares.foldLeft(cashInHand)((worth, kv) =>
        worth + (kv._2 * companies.get(kv._1).get.currentPrice))
